backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo[srv]>=4.7.2
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic[email]>=2.9.0
google-auth>=2.29.0
requests>=2.31.0
slowapi>=0.1.9
